+++ riscv_mpw/software/rv_lib/pim.h
@@ -8,9 +8,9 @@ void dump_buffer    (uint32_t source_addr, uint32_t size);
 
 /* PIM functions */
 void pim_erase      (uint32_t pulse_width, uint8_t pulse_count, uint8_t row);
-void pim_program    (uint32_t pulse_width, uint8_t pulse_count, uint8_t row, uint16_t col);
-void pim_read       (uint8_t row, uint16_t col);
-void pim_parallel   (uint32_t buffer_addr, uint8_t row, uint16_t col);
-void pim_rbr        (uint32_t buffer_addr, uint8_t row, uint16_t col);
+void pim_program    (uint32_t pulse_width, uint8_t pulse_count, uint8_t row, uint8_t col);
+void pim_read       (uint8_t row, uint8_t col);
+void pim_parallel   (uint32_t buffer_addr, uint8_t row, uint8_t col);
+void pim_rbr        (uint32_t buffer_addr, uint8_t row, uint8_t col);
 void pim_load       (uint32_t buffer_addr);
 void pim_debug      (uint32_t buffer_addr, uint8_t compute_mode);
